Time: 0 ms (100%), Space: 10.6 MB (71.97%) - LeetHub

+++ 1411-convert-binary-number-in-a-linked-list-to-integer/1411-convert-binary-number-in-a-linked-list-to-integer.cpp
@@ -12,7 +12,7 @@ class Solution {
 public:
     int getDecimalValue(ListNode* head) {
 
-        long long num = 0;
+        int num = 0;
         while(head){
             num = num << 1 | head->val;
             head = head -> next;
